Refactoring: extract code into new ServicesWorkflowHelper
