+++ lib/geet/services/open_pr.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative '../helpers/os_helper'
+require_relative '../helpers/services_workflow_helper'
 
 module Geet
   module Services
@@ -8,6 +9,7 @@ module Geet
     #
     class OpenPr
       include Geet::Helpers::OsHelper
+      include Geet::Helpers::ServicesWorkflowHelper
 
       DEFAULT_GIT_CLIENT = Geet::Utils::GitClient.new
 
@@ -22,23 +24,6 @@ module Geet
         pr = checked_find_branch_pr(merge_owner, merge_head)
         open_file_with_default_application(pr.link)
         pr
-      end
-
-      private
-
-      def find_merge_head
-        [@git_client.owner, @git_client.current_branch]
-      end
-
-      # Expect to find only one.
-      def checked_find_branch_pr(owner, head)
-        @out.puts "Finding PR with head (#{owner}:#{head})..."
-
-        prs = @repository.prs(owner: owner, head: head)
-
-        raise "Expected to find only one PR for the current branch; found: #{prs.size}" if prs.size != 1
-
-        prs[0]
       end
     end
   end
